pycon-backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "PyCon Italia backend"
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "=3.1.4"
django-environ = "=0.4.4"
social-auth-app-django = "^4.0.0"
django-model-utils = "=4.1.1"
pytz = "=2019.3"
django-timezone-field = "=4.0"
django-fsm = "=2.6.1"
psycopg2-binary = "=2.8.6"
amqp = "=2.5.1"
billiard = "=3.6.1.0"
celery = "=4.3.0"
kombu = "=4.6.5"
redis = "=3.3.11"
vine = "=1.3.0"
certifi = "=2019.9.11"
chardet = "=3.0.4"
idna = "=2.8"
requests = "=2.22.0"
urllib3 = "=1.25.7"
aniso8601 = "=8.0.0"
Pillow = "=6.2.1"
whitenoise = "=5.2.0"
django-ordered-model = "=3.3.0"
uWSGI = "=2.0.18"
django-storages = "=1.11.1"
sentry-sdk = "0.19.5"
boto3 = "^1.16"
pycountry = "^19.8"
countries = "^0.2.0"
django-ses = "^1.0.3"
hashids = "^1.2.0"
django-admin-views = {git = "https://github.com/estyxx/django-admin-views.git"}
django-autocomplete-light = "3.5.1"
dal_admin_filters = "^0.3.3"
django-queryinspect = "^1.1.0"
lxml = "^4.5.0"
unidecode = "^1.1.1"
jsonschema = "^3.2.0"
mypy_extensions = "^0.4.3"
graphql-core = "^3.1.0"
strawberry-graphql = "0.44.2"
Werkzeug = "^1.0.1"

[tool.poetry.dev-dependencies]
pylint = "=2.4.4"
pylint-django = "=2.0.13"
pytest = "6.2.1"
pytest-cov = "=2.10.1"
pytest-django = "=4.1.0"
pytest-pylint = "=0.18.0"
pytest-factoryboy = "=2.0.3"
factory-boy = "=3.1.0"
mypy = "^0.750"
pdbpp = "^0.10.2"
flake8 = "^3.7"
black = {version = "^19.10b0",allow-prereleases = true}
requests-mock = "^1.7"
pytest-mock = "^3.4"
freezegun = "^0.3.12"
pytest-cache = "^1.0"
bpython = "^0.18"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
